Refactor Player
The player and evaluator is started with identical data preparation. To
prevent both function use repeated operation, the player is refactored.

This commit introduces startMPA function, which is a bootstrap for the
player and evaluator. This function will prepare the data and call
either player or evaluator depending on user input.